originalCIdの定義をchartFormat/に移動
